Move survey logic to its model
Refactor for #1488
